personal/requirements.txt
django==1.9.4
djangorestframework
